+++ framework/meta/src/cmd/contract/generate_snippets/snippet_sc_functions_gen.rs
