+++ framework/meta/src/cmd/contract/generate_snippets/snippet_template_gen.rs
