+++ framework/meta/src/cmd/contract/generate_snippets/snippet_type_map.rs
