+++ framework/meta/src/cmd/contract/meta_abi.rs
@@ -1,7 +1,6 @@
 use std::{
     fs::{create_dir_all, File},
     io::Write,
-    process::{Command, Output},
 };
 
 use crate::abi_json::{serialize_abi_to_json, ContractAbiJson};
@@ -38,23 +37,6 @@ impl MetaConfig {
             return String::new();
         }
 
-        Command::new("git")
-            .args(["describe"])
-            .output()
-            .map(git_describe_process_output)
-            .unwrap_or_default()
-    }
-}
-
-fn git_describe_process_output(output: Output) -> String {
-    if output.status.success() {
-        let mut result = String::from_utf8(output.stdout).unwrap_or_default();
-        if result.ends_with('\n') {
-            // for some reason we get a trailing newline
-            let _ = result.pop();
-        }
-        result
-    } else {
-        String::new()
+        crate::tools::git_describe()
     }
 }
