+++ framework/meta/src/cmd/contract/output_contract/multi_contract_serde_stack_size.rs
