+++ framework/meta/src/cmd/contract/output_contract/output_contract_allocator.rs
