+++ framework/meta/src/cmd/contract/output_contract/wasm_clean.rs
