+++ framework/meta/src/cmd/contract/validate_abi.rs
