+++ framework/meta/src/cmd/standalone/scen_test_gen/stg_write.rs
