+++ framework/meta/src/cmd/standalone/upgrade/upgrade_0_39.rs
