+++ framework/meta/tests/multi_contract_test.rs
@@ -1,5 +1,7 @@
 use multiversx_sc::abi::{ContractAbi, EndpointAbi};
-use multiversx_sc_meta::output_contract::{MultiContractConfigSerde, OutputContractConfig};
+use multiversx_sc_meta::cmd::contract::output_contract::{
+    MultiContractConfigSerde, OutputContractConfig,
+};
 
 fn get_serialized_toml() -> MultiContractConfigSerde {
     toml::from_str(
